pyproject.toml
[project]
name = "ops-activity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.5.0",
    "marimo>=0.16.0",
    "polars>=1.33.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.13.1",
]
